requirements.txt
langchain
langchain-groq
beautifulsoup4
requests
python-docx
pdfplumber
python-dotenv
streamlit
